requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
chardet==3.0.4
discord.py==1.3.3
google==2.0.3
idna==2.9
multidict==4.7.6
pymongo==3.10.1
python-dotenv==0.13.0
soupsieve==2.0.1
websockets==8.1
yarl==1.4.2
